Add left bottom branching to rectangle knapsack problem
The rectangle kanpsack problem gained left-bottom branching, which is
a very common strategy in packing problems.

The implementation creates combined x-y positions, on which
left-bottom placement is simply min-min branching. A valid index
implies that the rectangle is used, while an index that is larger that
possible implies that it is not used.

Statically ordering based on value per area is used as tie-breaking.

In order to also force the rotation, the x-y values are doubled and a
bool is added as well indicating the rotation. Four variants are
available for which rotation to choose: standing, laying, input, and
flipped.